scss-project/src/app/features/testing/forms/forms.component.ts
import { Component, OnInit } from '@angular/core';
import { FormBuilder, FormGroup, Validators } from '@angular/forms';

@Component({
  selector: 'app-forms',
  templateUrl: './forms.component.html',
  styleUrls: ['./forms.component.scss']
})
export class FormsComponent implements OnInit {
  customerForm:any = FormGroup;
  constructor(private formBuilder: FormBuilder) { }

  ngOnInit(): void {
    this.customerForm = this.formBuilder.group({
      customerName: ['', Validators.required],
      customerAadhar: ['', Validators.required],
      address: ['', Validators.required],
      phoneNumber: ['', [Validators.required, Validators.pattern('^[0-9]*$')]],
      customerMail: ['', [Validators.required, Validators.email]],
      state: ['', Validators.required]
    });
}
onSubmit(): void {
  if (this.customerForm.invalid) {
    return;
  }
  // Submit form data
  console.log(this.customerForm.value);

  // Reset form
  this.customerForm.reset();
}
}
